home11/requirements.txt
allure-pytest==2.9.45
pytest==7.0.1
selenium==4.0.0
requests==2.26.0
pytest-xdist==2.5.0